backend/requirements.txt
fastapi
uvicorn[standard]
openai
elevenlabs
PyPDF2
python-dotenv
python-multipart
pydantic
websockets
opencv-python
numpy

# New for HireByte v2.0
sentence-transformers
scikit-learn
textblob